Revert "renaming old implementation to _wbx before adding the ported one"
This reverts commit c6690c8af946bf6bdb418d4ad1759288c79b9cb0.
