+++ src/BizHawk.Emulation.Cores/Consoles/Sony/PSP/PPSSPP.IDriveLight.cs
@@ -1,8 +1,8 @@
 using BizHawk.Emulation.Common;
 
-namespace BizHawk.Emulation.Consoles.Sony.PSP_WBX
+namespace BizHawk.Emulation.Consoles.Sony.PSP
 {
-	public partial class PPSSPP_WBX : IDriveLight
+	public partial class PPSSPP : IDriveLight
 	{
 		public bool DriveLightEnabled { get; private set; }
 		public bool DriveLightOn { get; private set; }
